+++ plugins/filetools/plugin.js
@@ -276,6 +276,7 @@
 		}
 
 		this.uploaded = 0;
+		this.uploadTotal = null;
 
 		this.status = 'created';
 
@@ -341,6 +342,23 @@
 	 *
 	 * @readonly
 	 * @property {Number} total
+	 */
+
+	/**
+	 * The total size of upload data in bytes.
+	 * If `xhr.upload` object is present this value will indicate total size of the request payload, not only the file
+	 * size itself. If `xhr.upload` object is not available and real upload size cannot be obtained - this value will
+	 * be equal to {@link #total}. It has null value until upload size is known.
+	 *
+	 * 		loader.on( 'update', function() {
+	 * 			// Wait till uploadTotal is present.
+	 * 			if ( loader.uploadTotal ) {
+	 * 				console.log( 'uploadTotal: ' + loader.uploadTotal );
+	 * 			}
+	 * 		});
+	 *
+	 * @readonly
+	 * @property {Number} uploadTotal
 	 */
 
 	/**
@@ -518,22 +536,43 @@
 				xhr.abort();
 			};
 
-			xhr.onabort = function() {
-				loader.changeStatus( 'abort' );
-			};
+			xhr.onerror = onError;
+			xhr.onabort = onAbort;
 
-			xhr.onerror = function() {
-				loader.message = loader.lang.filetools.networkError;
-				loader.changeStatus( 'error' );
-			};
+			// #13533 - When xhr.upload is present attach onprogress, onerror and onabort functions to get actual upload
+			// information.
+			if ( xhr.upload ) {
+				xhr.upload.onprogress = function( evt ) {
+					if ( evt.lengthComputable ) {
+						// Set uploadTotal with correct data.
+						if ( !loader.uploadTotal ) {
+							loader.uploadTotal = evt.total;
+						}
+						loader.uploaded = evt.loaded;
+						loader.update();
+					}
+				};
 
-			xhr.onprogress = function( evt ) {
-				loader.uploaded = evt.loaded;
+				xhr.upload.onerror = onError;
+				xhr.upload.onabort = onAbort;
+
+			} else {
+				// #13533 - If xhr.upload is not supported - fire update event anyway and set uploadTotal to file size.
+				loader.uploadTotal = loader.total;
 				loader.update();
-			};
+			}
 
 			xhr.onload = function() {
-				loader.uploaded = loader.total;
+				// #13433 - Call update at the end of the upload. When xhr.upload object is not supported there will be
+				// no update events fired during the whole process.
+				loader.update();
+
+				// #13433 - Check if loader was not aborted during last update.
+				if ( loader.status == 'abort' ) {
+					return;
+				}
+
+				loader.uploaded = loader.uploadTotal;
 
 				if ( xhr.status < 200 || xhr.status > 299 ) {
 					loader.message = loader.lang.filetools[ 'httpError' + xhr.status ];
@@ -563,6 +602,24 @@
 					}
 				}
 			};
+
+			function onError() {
+				// Prevent changing status twice, when HHR.error and XHR.upload.onerror could be called together.
+				if ( loader.status == 'error' ) {
+					return;
+				}
+
+				loader.message = loader.lang.filetools.networkError;
+				loader.changeStatus( 'error' );
+			}
+
+			function onAbort() {
+				// Prevent changing status twice, when HHR.onabort and XHR.upload.onabort could be called together.
+				if ( loader.status == 'abort' ) {
+					return;
+				}
+				loader.changeStatus( 'abort' );
+			}
 		},
 
 		/**